+++ src/emilybot/utils/inflect.py
@@ -8,7 +8,7 @@ def inflect(text: str) -> str:
     Example:
 
     >>> inflect("I saw no('duck', 5)")
-    "I saw 5 ducks"
+    'I saw 5 ducks'
     """
 
     return _inflect.inflect(text)  # type: ignore